+++ pandora-client-web/src/components/wardrobe/wardrobeItems.tsx
@@ -109,7 +109,7 @@ export function WardrobeItemManipulation({ className }: { className?: string; })
 	const itemLimit: number = target.type === 'character' ? ITEM_LIMIT_CHARACTER_WORN : ITEM_LIMIT_ROOM_INVENTORY;
 	const titlePrefix: string = target.type === 'character' ? 'Currently worn items' : 'Room inventory';
 
-	const title = `${ titlePrefix }, used: ${ itemCount } / ${ itemLimit } (${100 * itemCount / itemLimit}%)`;
+	const title = `${ titlePrefix }, used: ${ itemCount } / ${ itemLimit } (${ Math.ceil(100 * itemCount / itemLimit) }%)`;
 	const isRoomInventory = target.type === 'room' && currentFocus.container.length === 0;
 
 	const singleItemContainer = useMemo<boolean>(() => {
